# frontend/web/src/api/openapi.json
{
 "components": {
  "schemas": {
...
   "Checkout": {
    "description": "Everything the browser needs to hand the reader over to PAYUNi, and nothing secret.",
    "properties": {
     "amount": {
      "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
      "title": "Amount",
      "type": "string"
     },
     "currency": {
      "title": "Currency",
      "type": "string"
     },
     "fields": {
      "additionalProperties": {
       "type": "string"
      },
      "title": "Fields",
      "type": "object"
     },
     "mer_trade_no": {
      "title": "Mer Trade No",
      "type": "string"
     },
     "order_id": {
      "format": "uuid",
      "title": "Order Id",
      "type": "string"
     },
     "url": {
      "title": "Url",
      "type": "string"
     }
    },
    "required": [
     "order_id",
     "mer_trade_no",
     "amount",
     "currency",
     "url",
     "fields"
    ],
    "title": "Checkout",
    "type": "object"
   },
   "CheckoutRequest": {
    "properties": {
     "company": {
      "anyOf": [
       {
        "maxLength": 100,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Company"
     },
     "interval": {
      "default": "year",
      "enum": [
       "month",
       "year"
      ],
      "title": "Interval",
      "type": "string"
     },
     "lang": {
      "default": "zh-TW",
      "pattern": "^[a-z]{2}(-[A-Z][A-Za-z]{1,3})?$",
      "title": "Lang",
      "type": "string"
     }
    },
    "title": "CheckoutRequest",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "Offer": {
    "description": "What is for sale, for a page that has to name a price before anybody clicks.",
    "properties": {
     "amount": {
      "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
      "title": "Amount",
      "type": "string"
     },
     "available": {
      "default": true,
      "title": "Available",
      "type": "boolean"
     },
     "currency": {
      "title": "Currency",
      "type": "string"
     },
     "interval": {
      "title": "Interval",
      "type": "string"
     }
    },
    "required": [
     "amount",
     "currency",
     "interval"
    ],
    "title": "Offer",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "title": "Autora API",
  "version": "0.0.1"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/checkout": {
   "post": {
    "description": "Open an order and build the form that opens PAYUNi's page.\n\nSigning in comes first: the order records who a year is for, and a reader id is the only\nname this layer has for anybody (D-018). Nothing is granted here — the order is PENDING\nuntil PAYUNi says otherwise, even if the reader never comes back.\n\nClosed in P2 (D-231): refused first, so no order is ever opened, at any price.",
    "operationId": "start_checkout_api_checkout_post",
    "parameters": [
     {
      "in": "cookie",
      "name": "autora_reader",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Autora Reader"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CheckoutRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Checkout"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Start Checkout",
    "tags": [
     "payments"
    ]
   }
  },
  "/api/checkout/offer": {
   "get": {
    "description": "What a month or a year costs here. ``available`` is false when that one is not for sale\nyet, which is a fact about the site rather than an error — the page says \"soon\" instead of a\nprice.",
    "operationId": "get_offer_api_checkout_offer_get",
    "parameters": [
     {
      "in": "query",
      "name": "company",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "maxLength": 100,
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Company"
      }
     },
     {
      "in": "query",
      "name": "interval",
      "required": false,
      "schema": {
       "default": "year",
       "enum": [
        "month",
        "year"
       ],
       "title": "Interval",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Offer"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Offer",
    "tags": [
     "payments"
    ]
   }
  },
...
 }
}